# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tài liệu API cho BTL Nhóm 2",
        "description": "Đây là tài liệu API cho ứng dụng điểm danh/quản lý lịch dạy.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/schedules": {
            "get": {
                "tags": [
                    "Schedules (CRUD)"
                ],
                "summary": "Lấy danh sách Lịch học (Chưa triển khai)",
                "operationId": "getSchedulesList",
                "responses": {
                    "200": {
                        "description": "Chưa triển khai"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Schedules (CRUD)"
                ],
                "summary": "Tạo Lịch học (Chưa triển khai)",
                "operationId": "storeSchedule",
                "responses": {
                    "200": {
                        "description": "Chưa triển khai"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/schedules/{schedule}": {
            "get": {
                "tags": [
                    "Schedules (CRUD)"
                ],
                "summary": "Lấy 1 Lịch học (Chưa triển khai)",
                "operationId": "getScheduleById",
                "parameters": [
                    {
                        "name": "schedule",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chưa triển khai"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Schedules (CRUD)"
                ],
                "summary": "Cập nhật Lịch học (Chưa triển khai)",
                "operationId": "updateSchedule",
                "parameters": [
                    {
                        "name": "schedule",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chưa triển khai"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Schedules (CRUD)"
                ],
                "summary": "Xóa Lịch học (Chưa triển khai)",
                "operationId": "deleteSchedule",
                "parameters": [
                    {
                        "name": "schedule",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chưa triển khai"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{user}/schedules-by-date": {
            "get": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Lấy lịch dạy (theo ngày) của giáo viên (cho dropdown xin nghỉ)",
                "operationId": "getSchedulesByDateForTeacher",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID của giáo viên",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Ngày cần lấy lịch (Y-m-d)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-10-24"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "schedule_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "display_name": {
                                                "type": "string",
                                                "example": "Lập trình Web (IT1) - 1-2"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Lỗi validate ngày"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{user}/available-schedules-for-leave": {
            "get": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Lấy lịch dạy SẮP TỚI của giáo viên (để chọn khi xin nghỉ)",
                "operationId": "getAvailableSchedulesForLeave",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID của giáo viên",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "schedule_id": {
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "display_name": {
                                                "type": "string",
                                                "example": "24/10/2025 - 1-2 - Lập trình Web (IT1)"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Sử dụng Bearer token (Sanctum) sau khi đăng nhập",
                "name": "Bearer Token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}